+++ src/components/ciPipeline/Webhook/WebhookDetailsModal.tsx
@@ -13,7 +13,7 @@ import { Option } from '../../v2/common/ReactSelect.utils'
 import { components } from 'react-select'
 import { getUserRole, saveUser } from '../../userGroups/userGroup.service'
 import { ACCESS_TYPE_MAP, MODES } from '../../../config'
-import { createGeneratedAPIToken, getWebhookAPITokenList } from '../../apiTokens/service'
+import { createGeneratedAPIToken } from '../../apiTokens/service'
 import { useParams } from 'react-router-dom'
 import { getDateInMilliseconds } from '../../apiTokens/authorization.utils'
 import { ActionTypes, CreateUser, EntityTypes } from '../../userGroups/userGroups.types'
@@ -26,7 +26,7 @@ import {
     TOKEN_TAB_LIST,
 } from './webhook.utils'
 import { SchemaType, TabDetailsType, TokenListOptionsType, WebhookDetailsType, WebhookDetailType } from './types'
-import { executeWebhookAPI, getExternalCIConfig } from './webhook.service'
+import { executeWebhookAPI, getExternalCIConfig, getWebhookAPITokenList } from './webhook.service'
 import Tippy from '@tippyjs/react'
 import { toast } from 'react-toastify'
 import CodeEditor from '../../CodeEditor/CodeEditor'
