+++ src/components/ciPipeline/Webhook/webhook.service.ts
@@ -1,10 +1,9 @@
 import { Routes } from '../../../config'
 import { get } from '../../../services/api'
+import { ResponseType } from '../../../services/service.types'
 import { WebhookDetailsResponse, WebhookDetailsType, WebhookListResponse } from './types'
 
-export function getExternalCIList(
-    appId: number | string
-): Promise<WebhookListResponse> {
+export function getExternalCIList(appId: number | string): Promise<WebhookListResponse> {
     return get(`${Routes.EXTERNAL_CI_CONFIG}/${appId}`)
 }
 
@@ -13,6 +12,16 @@ export function getExternalCIConfig(
     webhookID: number | string,
 ): Promise<WebhookDetailsResponse> {
     return get(`${Routes.EXTERNAL_CI_CONFIG}/${appId}/${webhookID}`)
+}
+
+export function getWebhookAPITokenList(
+    projectName: string,
+    environmentName: string,
+    appName: string,
+): Promise<ResponseType> {
+    return get(
+        `${Routes.API_TOKEN_WEBHOOK}?projectName=${projectName}&environmentName=${environmentName}&appName=${appName}`,
+    )
 }
 
 export async function executeWebhookAPI(webhookUrl: string, token: string, data?: object): Promise<any> {
